finish building constants
